Add some nearly empty modules and a semi working logger
The nearly empty modules will shape a basic structure for the server. The
goal is to provide a very simple example server. Useful changes will be made
on top of that.

The logger is the only thing that will be delivered finished: It uses the
official `log` crate and provides nice and easy logging for the whole
project. Programming the logger is just about Rust, not database. The logger
is not finished yet, though.